+++ src/Command/Platformsh/Setup.php
@@ -66,11 +66,16 @@ class Setup extends DockerCommand
         // put the files in places
         $fs->mirror("{$this->getPayloadDir()}/platformsh/.platform", "{$this->projectPath}/.platform");
         $fs->copy("{$this->getPayloadDir()}/platformsh/.platform.app.yaml", "{$this->projectPath}/.platform.app.yaml");
-        $provisioningName   = $this->projectConfiguration->get('provisionning.folder_name');
+        $provisioningName   = $this->projectConfiguration->get('provisioning.folder_name');
         $provisioningFolder = "{$this->projectPath}/{$provisioningName}";
         $fs->copy(
             "{$this->getPayloadDir()}/platformsh/getmysqlcredentials.php",
             "{$provisioningFolder}/platformsh/getmysqlcredentials.php",
+            true
+        );
+        $fs->copy(
+            "{$this->getPayloadDir()}/platformsh/clearmemcache.php",
+            "{$provisioningFolder}/platformsh/clearmemcache.php",
             true
         );
         $fs->copy(
@@ -87,6 +92,7 @@ class Setup extends DockerCommand
             "\t<comment>git commit -m \"Integration Platform.sh\"</comment>\n".
             "\t<comment>git push platform {branchname}</comment>\n"
         );
+
         $this->postAction();
     }
 }