+++ src/bootstrap.php
@@ -10,12 +10,12 @@ if (!file_exists($file)) {
 
 if (!file_exists($file)) {
     echo 'You must set up the project dependencies, run the following commands:'.PHP_EOL.
-         'make install'.PHP_EOL;
+         'composer install'.PHP_EOL;
     exit(1);
 }
 
 include $file;
 
-if (@date_default_timezone_get() != '') {
+if ('' != @date_default_timezone_get()) {
     date_default_timezone_set(getDefaultTimeZone());
 }